feat: Add isBusy field to Consultant model
This commit adds a new field `isBusy` to the `Consultant` model in the `schema.prisma` file. The `isBusy` field is of type `Boolean` and is set to `false` by default. This field will be used to track the availability of a consultant for new projects.

The changes also include updating the `app.module.ts` file to import and use the `UsersController`, `UsersService`, and `PrismaService` in the `AppModule`.

These changes were made to enhance the functionality of the application and improve the management of consultants' availability.